Refactor to fix nested relative paths.
This is a significant refactor that improves support for relative paths
that are nested more than a single layer deep in the dependency tree.

Instead of copying directories into the cache directory,
packages that already exist on the file system are referenced in-place,
and not cached.

I chose this approach because moving packages into the cache directory breaks
relative paths that are nested more than one layer deep in the dependency tree,
unless those dependencies are located within the original package directory
(however, even this case would break if we start removing all files that don't
belong to a package before caching).

Currently, not even the hash of path type dependencies is cached, if
those dependencies point to a directory.  I chose this so that it is
easier to detect when the hash of dependencies must be updated in
build.zig.zon.

For example, imagine a vendored dependency that is referred to
by a relative path. If the hash of the package is changed, for example if a
patch is applied during development, then the package could change without
the build system detecting that the hash must be updated in build.zig.zon.
Subsequent users of the code would then encouter a build error when
trying to build for the first time. A similar situation could occur
if someone directly edited files in the global cache, but I feel that this
case is more of a concern because the user is more likely to edit files
referred to by path (as in the case of vendored dependencies) than files
fetched from the internet.

The downside of this choice is that the hash for packages referred
to by path is re-computed on every build (Note that this is not the case
if the path refers to a file that must be unpacked, such as a .tar.gz).